Remove automatic header IDs
The newer MDX Acorn parser throws on the custom ID syntax. It is common and conventional syntax, but non-standard.

For predictable IDs, the alternative syntax isn't that bad: manually add an `<a>` before the heading.

For the convenience of a link icon appearing on hover, this is a loss of functionality. But I don't want the automatic links interfering with manually maintained ones.